+++ ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/Service/UniversityClassServiceImpl.java
@@ -6,18 +6,20 @@ import com.chk.ubbprotool.ubbprotool.Model.Teacher;
 import com.chk.ubbprotool.ubbprotool.Model.UniversityClass;
 import com.chk.ubbprotool.ubbprotool.Repository.*;
 import com.chk.ubbprotool.ubbprotool.dto.StudentDTO;
+import com.chk.ubbprotool.ubbprotool.Repository.StudentRepository;
+import com.chk.ubbprotool.ubbprotool.Repository.SubgroupRepository;
+import com.chk.ubbprotool.ubbprotool.Repository.UniversityClassRepository;
+import com.chk.ubbprotool.ubbprotool.Repository.WeeksRepository;
 import com.chk.ubbprotool.ubbprotool.dto.UniversityClassDTO;
 import com.chk.ubbprotool.ubbprotool.mapper.UniversityClassMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.sql.Date;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 
 @Component
 @ComponentScan(basePackages = "com.chk.ubbprotool.ubbprotool")
@@ -76,14 +78,14 @@ public class UniversityClassServiceImpl implements UniversityClassService{
 
     @Override
     @Transactional
-    public UniversityClassDTO findById(int id) {
+    public UniversityClassDTO findById(Long id) {
         UniversityClass universityClass = universityClassRepository.findById(id);
         return universityClassMapper.toDTO(universityClass);
     }
 
     @Override
     @Transactional
-    public List<UniversityClassDTO> getClassesForStudent(int studentId, Date date) {
+    public List<UniversityClassDTO> getClassesForStudent(Long studentId, Date date) {
 
         Student student = studentRepository.findById(studentId);
