+++ ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/dto/UniversityClassDTO.java
@@ -1,19 +1,10 @@
 package com.chk.ubbprotool.ubbprotool.dto;
 
-import com.chk.ubbprotool.ubbprotool.Model.Course;
-import com.chk.ubbprotool.ubbprotool.Model.Subgroup;
-import com.chk.ubbprotool.ubbprotool.Model.Teacher;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.NonNull;
 import org.springframework.boot.jackson.JsonComponent;
 import org.springframework.stereotype.Component;
-
-import javax.persistence.Column;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import java.sql.Time;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -40,5 +31,5 @@ public class UniversityClassDTO {
 
     private int courseId;
 
-    private int subgroupId;
+    private Long subgroupId;
 }
